update pre V6.0.0
