+++ config/initialCommand.js
@@ -1,0 +1,9 @@
+const {selesaiRestart} = require('../lib/restartServer');
+
+function initialCommand() {
+    selesaiRestart();
+}
+    
+module.exports ={
+    initialCommand
+}
